requirements.txt
python-dotenv
selenium
beautifulsoup4
psycopg2-binary
mailerlite
psycopg
requests
fastapi
uvicorn
